+++ CHANGELOG.md
@@ -24,7 +24,9 @@ _ `_optional` subpackage for managing optional dependencies
 - `ContinuousCardinalityConstraint` class and corresponding uniform sampling mechanism
 - `register_hooks` utility enabling user-defined augmentation of arbitrary callables
 - Polars expressions for `DiscreteSumConstraint`, `DiscreteProductConstraint`, 
-  `DiscreteExcludeConstraint`, `DiscreteLinkedParametersConstraint` and `DiscreteNoLabelDuplicatesConstraint`
+  `DiscreteExcludeConstraint`, `DiscreteLinkedParametersConstraint` and 
+  `DiscreteNoLabelDuplicatesConstraint`
+- Discrete search space Cartesian product is created lazily via Polars
 
 ### Changed
 - Passing an `Objective` to `Campaign` is now optional